requirements.txt
python-telegram-bot>=20.6
sentence-transformers>=2.2.2
transformers>=4.40.0
torch
numpy
